fix: refactoring structure
