+++ src/components/ui/Form/index.ts
@@ -1,0 +1,9 @@
+export * from './Checkbox/Checkbox';
+export * from './Input/Input';
+export * from './InputCheckbox/InputCheckbox';
+export * from './InputDate/InputDate';
+export * from './InputFile/InputFile';
+export * from './Label/Label';
+export * from './Radio/Radio';
+export * from './Search/Search';
+export * from './Select/Select';
