+++ src/components/ui/index.ts
@@ -1,0 +1,3 @@
+export * from './Alert/Alert';
+export * from './Button/Button';
+export * from './TabLink/TabLink';
