+++ src/pages/About/About.tsx
@@ -1,6 +1,6 @@
 import { Component } from 'react';
 
-class About extends Component {
+export class About extends Component {
   render() {
     return (
       <>
@@ -10,5 +10,3 @@ class About extends Component {
     );
   }
 }
-
-export default About;
