+++ src/pages/index.ts
@@ -1,0 +1,3 @@
+export * from './About/About';
+export * from './Creation/Creation';
+export * from './Home/Home';
